backend/requirements.txt
Flask
Flask-Login
bcrypt
psycopg2-binary
python-dotenv
PyCryptodome
Pillow
email-validator
flask-cors

